Junits added

+++ src/main/java/com/srikanth/dao/EmployeeDao.java
@@ -1,6 +1,6 @@
 package com.srikanth.dao;
 
-import com.srikanth.exception.EmployeeManagementException;
+import com.srikanth.exception.EmployeeNotFoundException;
 import com.srikanth.model.Employee;
 import com.srikanth.util.HibernateSessionUtil;
 import com.srikanth.util.JerseyAppConstants;
@@ -11,6 +11,7 @@ import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 
+import javax.persistence.PersistenceException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -21,7 +22,7 @@ public class EmployeeDao {
     private static final Logger logger = Logger.getLogger(EmployeeDao.class);
 
 
-    public Employee addEmployee(Employee employee) throws EmployeeManagementException {
+    public Employee addEmployee(Employee employee) {
         logger.info("Persisting employee to database");
         Session session = null;
         Transaction transaction = null;
@@ -32,10 +33,10 @@ public class EmployeeDao {
             employee.setEmployeeId(UUID.randomUUID().toString().split("-")[0]);
             session.persist(employee);
             transaction.commit();
-        } catch (HibernateException ex) {
+        } catch (PersistenceException ex) {
             logger.error("Failed saving employee. Rolling back the transaction");
             transaction.rollback();
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
+            throw ex;
         } finally {
             if (session != null) {
                 session.close();
@@ -44,7 +45,7 @@ public class EmployeeDao {
         return employee;
     }
 
-    public Employee deleteEmployee(Employee employee) throws EmployeeManagementException {
+    public Employee deleteEmployee(Employee employee) {
         logger.info("Deleting employee from database");
         Session session = null;
         Transaction transaction = null;
@@ -57,7 +58,7 @@ public class EmployeeDao {
         } catch (HibernateException ex) {
             logger.error("Failed deleting employee. Rolling back the transaction");
             transaction.rollback();
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
+            throw ex;
         } finally {
             if (session != null) {
                 session.close();
@@ -66,7 +67,7 @@ public class EmployeeDao {
         return employee;
     }
 
-    public Employee deleteEmployee(String employeeId) throws EmployeeManagementException {
+    public Employee deleteEmployee(String employeeId) {
         logger.info("Deleting employee from database");
         Session session = null;
         Transaction transaction = null;
@@ -75,13 +76,15 @@ public class EmployeeDao {
             SessionFactory sessionFactory = HibernateSessionUtil.getSessionFactory();
             session = sessionFactory.openSession();
             transaction = session.beginTransaction();
-            employeeToDel = (Employee) session.load(Employee.class, employeeId);
-            session.delete(employeeToDel);
-            transaction.commit();
+            employeeToDel = session.get(Employee.class, employeeId);
+            if (employeeToDel != null) {
+                session.delete(employeeToDel);
+                transaction.commit();
+            }
         } catch (HibernateException ex) {
             logger.error("Failed deleting employee. Rolling back the transaction");
             transaction.rollback();
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
+            throw ex;
         } finally {
             if (session != null) {
                 session.close();
@@ -90,7 +93,7 @@ public class EmployeeDao {
         return employeeToDel;
     }
 
-    public Collection<Employee> getEmployees() throws EmployeeManagementException {
+    public Collection<Employee> getEmployees() {
         logger.info("Fetching all employees");
         List<Employee> employeeList = new ArrayList<>();
         Session session = null;
@@ -98,9 +101,6 @@ public class EmployeeDao {
             SessionFactory sessionFactory = HibernateSessionUtil.getSessionFactory();
             session = sessionFactory.openSession();
             employeeList = session.createQuery("from Employee").list();
-        } catch (HibernateException ex) {
-            logger.error("Failed fetching employees list\n" + ex);
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
         } finally {
             if (session != null) {
                 session.close();
@@ -109,7 +109,7 @@ public class EmployeeDao {
         return employeeList;
     }
 
-    public Collection<Employee> getEmployee(String username, String password) throws EmployeeManagementException {
+    public Collection<Employee> getEmployee(String username, String password) {
         logger.info("Fetching employee with username " + username);
         List<Employee> employeeList = new ArrayList<>();
         Session session = null;
@@ -120,9 +120,6 @@ public class EmployeeDao {
                     .setParameter("username", username)
                     .setParameter("password", password);
             employeeList = query.list();
-        } catch (HibernateException ex) {
-            logger.error("Failed fetching employees list\n" + ex);
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
         } finally {
             if (session != null) {
                 session.close();
@@ -132,7 +129,7 @@ public class EmployeeDao {
     }
 
 
-    public Employee getEmployee(String id) throws EmployeeManagementException {
+    public Employee getEmployee(String id) {
         logger.info("Fetching Employee using employee id " + id);
         Employee employee = null;
         Session session = null;
@@ -140,9 +137,7 @@ public class EmployeeDao {
             SessionFactory sessionFactory = HibernateSessionUtil.getSessionFactory();
             session = sessionFactory.openSession();
             employee = session.get(Employee.class, id);
-        } catch (HibernateException ex) {
-            logger.error("Failed fetching employee\n" + ex);
-            throw new EmployeeManagementException("Exception during database transaction" + ex);
+
         } finally {
             if (session != null) {
                 session.close();
